StepNumber rewritten to TS
