+++ src/components/StepNumber/StepNumber.mdx
@@ -1,5 +1,5 @@
 import { Playground, Props } from 'docz'
-import StepNumber from '@components/StepNumber'
+import { StepNumber } from '@components/StepNumber'
 
 # StepNumber
