+++ src/components/StepNumber/index.ts
@@ -1,0 +1,1 @@
+export * from './StepNumber'
