+++ src/components/index.ts
@@ -87,7 +87,7 @@ export { default as Spreader } from './Spreader'
 export { default as StatsBadge } from './StatsBadge'
 export { default as StatusDropdown } from './StatusDropdown'
 export { default as StatusIcon } from './StatusIcon'
-export { default as StepNumber } from './StepNumber'
+export * from './StepNumber'
 export { default as Steps } from './Steps'
 export { default as Tab } from './Tabs/Tab'
 export { default as TabList } from './Tabs/TabList'
